refactor(x/ts): rename string package to strings
